requirements.txt
mysql-connector-python
sendgrid
schedule
selenium
python-dotenv